requirements.txt
python-dotenv

langchain-core
langchain-community
langchain-chroma
langchain-ollama
langchain-text-splitters

chromadb
rank-bm25

sentence-transformers
torch
transformers

unstructured[pdf]
pillow
pypdf